Movido os comandos para estrutura propria
