+++ src/api/carolApi.ts
@@ -1,11 +1,10 @@
 import * as vscode from "vscode";
 
-import { TDitoConfig, getDitoConfiguration, getDitoUser, setDitoReady, setDitoUser } from "../config";
+import { TDitoConfig, getDitoConfiguration, getDitoUser, setDitoUser } from "../config";
 import { fetch, Response } from "undici";
 import { capitalize } from "../util";
 import { Completion, CompletionResponse, IaAbstractApi, IaApiInterface, TypifyResponse } from "./interfaceApi";
-import { PREFIX_DITO, logger } from "../logger";
-import { updateContextKey } from "../extension";
+import { logger } from "../logger";
 
 export class CarolApi extends IaAbstractApi implements IaApiInterface {
     // prefixo _ indica envolvidas com a API CAROL
@@ -330,307 +329,9 @@ export class CarolApi extends IaAbstractApi implements IaApiInterface {
      * 
      */
     register(context: vscode.ExtensionContext): void {
-
-        /**
-         * Registers a command with VS Code to prompt the user to login.
-         * 
-         * Checks if an API token is already stored, and attempts auto-login if so.
-         * Otherwise, prompts the user to enter their API token or username. 
-         * Validates the login and stores the token if successful.
-         * 
-         * @param args - First arg is a boolean to skip auto-login if true.
-        */
-        //TODO: Para identificação do usuário, implementar usando AuthProvider
-        //const authProvider = new AuthProvider(initialConfig)
-        //await authProvider.init()
-
-        context.subscriptions.push(vscode.commands.registerCommand('tds-dito.login', async (...args) => {
-            let apiToken = await context.secrets.get('apiToken');
-
-            if (apiToken !== undefined) {
-                this.start(apiToken).then(async (value: boolean) => {
-                    if (await this.login()) {
-                        logger.info('Logged in successfully');
-                        return;
-                    }
-                });
-            }
-
-            if (args.length > 0) {
-                if (args[0]) { //indica que login foi acionado automaticamente
-                    return;
-                }
-            }
-
-            const input = await vscode.window.showInputBox({
-                prompt: 'Please enter your API token or @your name):',
-                placeHolder: 'Your token goes here ...'
-            });
-            if (input !== undefined) {
-                if (await this.start(input)) {
-                    if (await this.login()) {
-                        await context.secrets.store('apiToken', input);
-                        vscode.window.showInformationMessage(`${PREFIX_DITO} Logged in successfully`);
-                    } else {
-                        await context.secrets.delete('apiToken');
-                        vscode.window.showErrorMessage(`${PREFIX_DITO} Login failure`);
-                    }
-
-                    this.chat.checkUser("");
-                }
-            }
-        }));
-
-        /**
-         * Registers a command to log out the user by deleting the stored API token.
-         * Logs the user out, deletes the stored API token and shows an informational message.
-         */
-        context.subscriptions.push(vscode.commands.registerCommand('tds-dito.logout', async (...args) => {
-            this.chat.logout();
-            this.logout();
-            await context.secrets.delete('apiToken');
-            vscode.window.showInformationMessage(`${PREFIX_DITO} Logged out`);
-            this.chat.checkUser("");
-        }));
-
-        /**
-         * Registers a health check command that checks the health of the Dito service. 
-         * Shows a detailed error message if the health check fails.
-         * On success, logs in the user automatically.
-        */
-        context.subscriptions.push(vscode.commands.registerCommand('tds-dito.health', async (...args) => {
-            let detail: boolean = true;
-            const messageId = this.chat.dito("Verificando disponibilidade do serviço.");
-
-            return new Promise((resolve, reject) => {
-                if (args.length > 0) {
-                    if (!args[0]) { //solicitando verificação sem  detalhes
-                        detail = false;
-                    }
-                }
-
-                this.checkHealth(detail).then((error: any) => {
-                    updateContextKey("readyForUse", error === undefined);
-                    setDitoReady(error === undefined);
-
-                    if (error !== undefined) {
-                        const message: string = `Desculpe, estou com dificuldades técnicas. ${this.chat.commandText("health")}`;
-                        this.chat.dito(message, messageId);
-                        vscode.window.showErrorMessage(`${PREFIX_DITO} ${message}`);
-
-                        if (error.message.includes("502: Bad Gateway")) {
-                            const parts: string = error.message.split("\n");
-                            //TODO: Motta
-                            this.chat.ditoInfo(parts[1]);
-                        }
-
-                        if (detail) {
-                            this.chat.ditoInfo(JSON.stringify(error, undefined, 2));
-                        }
-                    } else {
-                        vscode.commands.executeCommand("tds-dito.login", true).then(() => {
-                            this.chat.checkUser(messageId);
-                        });
-                    }
-                })
-            });
-        }));
-
-        //o que o usuário descrever vai vim veio argumento
-        context.subscriptions.push(vscode.commands.registerTextEditorCommand('tds-dito.generate-code', () => {
-            const text: string = "Gerar código para varrer um array";
-            this.generateCode(text);
-        }));
-
-        /**
-         * Registers a text editor command to explain the code under the cursor or selection. 
-         * Checks if there is an active text editor, gets the current selection or line under cursor, 
-         * extracts the code to explain, sends it to the explainCode() method, 
-         * and prints the explanation to the chat window.
-         */
-        context.subscriptions.push(vscode.commands.registerTextEditorCommand('tds-dito.explain', () => {
-            const editor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
-            let codeToExplain: string = "";
-
-            if (editor !== undefined) {
-                const selection: vscode.Selection = editor.selection;
-                let whatExplain: string = "";
-
-                if (selection && !selection.isEmpty) {
-                    const selectionRange: vscode.Range = new vscode.Range(selection.start.line, selection.start.character, selection.end.line, selection.end.character);
-                    codeToExplain = editor.document.getText(selectionRange);
-                    whatExplain = this.chat.linkToSource(editor.document.uri, selectionRange);
-
-                } else {
-                    const curPos: vscode.Position = selection.start;
-                    const contentLine: string = editor.document.lineAt(curPos.line).text;
-
-                    whatExplain = this.chat.linkToSource(editor.document.uri, curPos.line);
-                    codeToExplain = contentLine.trim();
-                }
-
-                if (codeToExplain.length > 0) {
-                    const messageId: string = this.chat.dito(
-                        `Explicando o código ${whatExplain}`
-                    );
-
-                    return this.explainCode(codeToExplain).then((value: string) => {
-                        if (getDitoConfiguration().clearBeforeExplain) {
-                            this.chat.dito("clear");
-                        }
-                        this.chat.dito(value, messageId);
-                    });
-                } else {
-                    this.chat.ditoWarning("Não consegui identificar um código para explica-lo.");
-                }
-            } else {
-                this.chat.ditoWarning("Editor corrente não é valido para essa operação.");
-            }
-        }));
-
-        /**
-         * Registers a text editor command to explain the word under the cursor selection. 
-         * Gets the current active text editor, then gets the word range at the cursor position. 
-         * Sends the word to be explained to the chatbot.
-         * Displays the explanation in the chat window.
-        */
-        context.subscriptions.push(vscode.commands.registerTextEditorCommand('tds-dito.explain-word', () => {
-            const editor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
-
-            if (editor !== undefined) {
-                const selection: vscode.Selection = editor.selection;
-                const selectionRange: vscode.Range | undefined = editor.document.getWordRangeAtPosition(selection.start);
-
-                if (selectionRange !== undefined) {
-                    let wordToExplain: string = editor.document.getText(selectionRange).trim();
-                    let whatExplain = this.chat.linkToSource(editor.document.uri, selectionRange);
-
-                    if (wordToExplain.length > 0) {
-                        //const workspaceFolder: vscode.WorkspaceFolder | undefined = vscode.workspace.getWorkspaceFolder(editor.document.uri);
-
-                        const messageId: string = this.chat.dito(
-                            `Explicando palavra ${whatExplain}`
-                        );
-
-                        return this.explainCode(wordToExplain).then((value: string) => {
-                            if (getDitoConfiguration().clearBeforeExplain) {
-                                this.chat.dito("clear");
-                            }
-                            this.chat.dito(value, messageId);
-                        });
-                    } else {
-                        this.chat.ditoWarning("Não consegui identificar uma palavra para explica-la.");
-                    }
-                } else {
-                    this.chat.ditoWarning("Não consegui identificar uma palavra para explica-la.");
-                }
-            } else {
-                this.chat.ditoWarning("Editor corrente não é valido para essa operação.");
-            }
-        }));
-
-        /**
-         * Registers a command to infer types for a selected function in the active text editor. 
-         * Finds the enclosing function based on the cursor position, extracts the function code, and sends it to an API to infer types.
-         * Displays the inferred types in the chat window.
-        */
-        context.subscriptions.push(vscode.commands.registerCommand('tds-dito.typify', async (...args) => {
-            const editor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
-            let codeToTypify: string = "";
-
-            if (editor !== undefined) {
-                if (getDitoConfiguration().clearBeforeExplain) {
-                    this.chat.dito("clear");
-                }
-
-                const selection: vscode.Selection = editor.selection;
-                const function_re: RegExp = /(function|method(...)class)\s*(\w+)/i
-                const return_re: RegExp = /^\s*(Return|EndClass)/i
-                const curPos: vscode.Position = selection.start;
-                let whatExplain: string = "";
-                let curLine = curPos.line;
-                let startFunction: vscode.Position | undefined = undefined;
-                let endFunction: vscode.Position | undefined = undefined;
-
-                //começo da função
-                while ((curLine > 0) && (!startFunction)) {
-                    const lineStart = new vscode.Position(curLine - 1, 0);
-                    const curLineStart = new vscode.Position(lineStart.line, 0);
-                    const nextLineStart = new vscode.Position(lineStart.line + 1, 0);
-                    const rangeWithFirstCharOfNextLine = new vscode.Range(curLineStart, nextLineStart);
-                    const contentWithFirstCharOfNextLine = editor.document.getText(rangeWithFirstCharOfNextLine);
-
-                    if (contentWithFirstCharOfNextLine.match(function_re)) {
-                        startFunction = new vscode.Position(curLine, 0);
-                    }
-
-                    curLine--;
-                }
-
-                curLine = curPos.line;
-
-                while ((curLine < editor.document.lineCount) && (!endFunction)) {
-                    const lineStart = new vscode.Position(curLine + 1, 0);
-                    const curLineStart = new vscode.Position(lineStart.line, 0);
-                    const nextLineStart = new vscode.Position(lineStart.line + 1, 0);
-                    const rangeWithFirstCharOfNextLine = new vscode.Range(curLineStart, nextLineStart);
-                    const contentWithFirstCharOfNextLine = editor.document.getText(rangeWithFirstCharOfNextLine);
-                    const matches = contentWithFirstCharOfNextLine.match(return_re);
-
-                    if (matches) {
-                        endFunction = new vscode.Position(curLine, 0);
-                    }
-
-                    curLine++;
-                }
-
-                if (startFunction) {
-                    if (!endFunction) {
-                        endFunction = new vscode.Position(editor.document.lineCount - 1, 0);
-                    }
-
-                    const rangeForTypify = new vscode.Range(startFunction, endFunction);
-                    codeToTypify = editor.document.getText(rangeForTypify);
-
-                    if (codeToTypify.length > 0) {
-                        const rangeBlock = new vscode.Range(startFunction, endFunction);
-
-                        whatExplain = this.chat.linkToSource(editor.document.uri, rangeBlock);
-
-                        const messageId: string = this.chat.dito(
-                            `Tipificando o código ${whatExplain}`
-                        );
-
-                        return this.typify(codeToTypify).then((response: TypifyResponse) => {
-                            let text: string[] = [];
-
-                            if (response !== undefined && response.types !== undefined && response.types.length) {
-                                for (const varType of response.types) {
-                                    text.push(`- **${varType.var}** as **${varType.type}** ${this.chat.commandText("update")}`);
-                                }
-                                this.chat.dito(text, messageId);
-
-                                //const edit = new vscode.WorkspaceEdit();
-                                //edit.insert(YOUR_URI, rangeForTypify.start, "TEST");
-                                //let success = await vscode.workspace.applyEdit(edit);
-                                const ss: vscode.SnippetString = new vscode.SnippetString;
-                                ss.appendText("TEST");
-                                editor.insertSnippet(ss,rangeForTypify);
-
-                            } else {
-                                this.chat.dito("Desculpe, não consegui fazer a tipificação por conta de um problema interno.", messageId);
-                            }
-                        });
-                    }
-                } else {
-                    this.chat.ditoWarning([
-                        "Não consegui identificar uma função/método para tipificar.",
-                        "Experimente posicionar o cursor em outra linha da implementação."
-                    ]);
-                }
-            } else {
-                this.chat.ditoWarning("Editor corrente não é valido para essa operação.");
-            }
-        }));
+      
+        //Os registros de comandos devem ficar em uma estrutura propria.
+        //Seguir o exemplo da pasta src/commands/IA
+              
     }
 }
