+++ src/commands/chat/cmdClear.ts
@@ -1,0 +1,12 @@
+import * as vscode from "vscode";
+import { ChatApi } from '../../api/chatApi';
+
+
+export function registerClear(context: vscode.ExtensionContext, chatApi: ChatApi): void {
+
+    const clear = vscode.commands.registerCommand('tds-dito.clear', async () => {
+        chatApi.user("clear", true);
+    });
+    context.subscriptions.push(clear);
+
+}
